feat: dashboard

+++ src/Backend/Pliromi.Domain/Entities/Transaction.cs
@@ -3,7 +3,7 @@ namespace Pliromi.Domain.Entities;
 public class Transaction
 {
 	public Guid Id { get;  set; }
-	public decimal Value { get; set; }
+	public int Value { get; set; }
 	
 	public Guid SenderId { get;  set; }
 	public User Sender { get; set; } = null!;
